+++ bridges/interfaces/cli/__init__.py
@@ -1,0 +1,9 @@
+"""
+bridges.interfaces.cli
+
+CLI interface package for bridges framework.
+"""
+
+from .core import CLI
+
+__all__ = ["CLI"]